requirements.txt
pandas
sentence_transformers
torch
streamlit-aggrid